requirements.txt
# Core
pandas>=2.2.2
numpy>=1.26.4
pyarrow>=15.0.0
requests>=2.31.0
python-dotenv>=1.0.1

# Database
SQLAlchemy>=2.0.30
pymysql>=1.1.0

# Vector DB + AI
chromadb>=0.5.3
openai>=1.40.0

# API
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0

# UI
streamlit>=1.35.0
